Improve detection of Tcl and Tk under Darwin
Make use of the tclConfig.sh and tkConfig.sh files to detect Tcl and Tk
headers and libraries locations and compilation flags.

Removed hardcoded search path for headers and libs, since all the
information is included in the *Config.sh. A search list containing
various default places for *Config.sh is still retained.

A tricky part is added to find the proper -R flag to pass to libtool
(useless on Darwin but mandatory on Linux; NetBSD is somewhere in
between...)