Build eltclsh as a Tcl extension that can be loaded in an existing interpreter
eltclsh code was already properly decoupled, so providing eltclsh as a
package was mostly a matter of compiling things separately.

The eltclsh and elwish interpreters are now basic Tcl interpreters
that require the eltclsh package. Note that is you "package require
eltclsh" from a vanilla Tcl interpreter, you have to call the
'interactive' command to activate editline facilities.

A notable detail is that the -module option of libtool cannot be used
to compile the package: Tcl on Darwin expects its packages to be
".dylib" files (I think they should be .so and libtool is right here).
But using libtool without -module seems to work on linux and darwin and
I did not notice any difference between the two versions.