Initialize uninitialized internal data in the eltclsh package
When a/ eltclsh is loaded as a package in an existing tclsh interpreter and b/
the el::gets function is called and c/ no previous call to the 'interactive'
procedure was ever issued, private data happen to be uninitialized, making the
completion generation function segfault whenever tab was hit.

Kindly reported by Russell Treleaven.