Add a variable to limit the number of columns in completions output
el::maxCols can be set to a positive integer to limit the number of columns
when displaying completion results. For example, if set to 1, completion
results will be displayed one per line. Defaults to 0, which means automatic
calculation based on the terminal window size.

Slightly modified version of a patch kindly provided by Russell Treleaven.