Fix potential segfault when printing interactive prompt
Don't use multibyte (or latin1, for that matter) characters in default
prompt.

Recent editline tries converts them to wide chars but returns a NULL
pointer in case of failure, which is not tested aftewards, leading to a
segfault. Since the conversion depends on a lot of unknowns (current
locale, in particular), better stick to ascii for default prompts.

Fixes #82